Fixed Prismatic Joint, Added Group Controller
Previously the sliding joint that connects the leg to the test stand was locked to be stationary. This joint is supposed to slide up and down freely to be able to test different leg behavior. Through changing the axis of movement, and increasing the allowed bounds of movement, desired behavior was achieved.

Added a Group Joint Trajectory Controller - all joints can now be controlled from a single ROS message, which makes it a little clearer/easier to control. Furthermore, this interface better suits following a desired trajectory, which is more in line with how we want to plan goit.